+++ src/components/island/Island.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import Box, { pickBoxProps } from '../box';
 import cx from 'classnames';
 import theme from './theme.css';
-import { elementIsDark } from '../utils/utils';
 
 const SIZES = {
   small: 3,
@@ -12,10 +11,9 @@ const SIZES = {
 };
 
 const Island = (props) => {
-  const { children, className, color, dark, size, ...others } = props;
+  const { children, className, color, size, ...others } = props;
 
   const classNames = cx(theme[color], className);
-  const isDark = elementIsDark(color, dark);
   const boxProps = pickBoxProps(others);
 
   return (
@@ -25,7 +23,7 @@ const Island = (props) => {
       backgroundTint={color === 'neutral' ? 'light' : 'lightest'}
       borderRadius="rounded"
       borderColor={color === 'white' ? 'neutral' : color}
-      borderTint={isDark ? 'dark' : 'normal'}
+      borderTint="normal"
       borderBottomWidth={1}
       borderLeftWidth={1}
       borderRightWidth={1}
@@ -43,7 +41,6 @@ Island.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
   color: PropTypes.oneOf(['neutral', 'mint', 'violet', 'ruby', 'gold', 'aqua', 'white']),
-  dark: PropTypes.bool,
   size: PropTypes.oneOf(Object.keys(SIZES)),
 };