+++ src/components/pagination/pagination.stories.js
@@ -7,20 +7,7 @@ export default {
   title: addStoryInGroup(MID_LEVEL_BLOCKS, 'Pagination'),
 };
 
-export const Full = ({ inverse, ...args }) => (
-  <Pagination inverse={inverse} {...args}>
-    {({ number, text, isActive, ...others }) => {
-      return <Button level="link" label={text} disabled={isActive} inverse={inverse} size="small" {...others} />;
-    }}
-  </Pagination>
-);
-
-Full.args = {
-  currentPage: 3,
-  numPages: 21,
-};
-
-export const Compact = (args) => (
+export const DefaultStory = (args) => (
   <Pagination {...args}>
     {({ number, text, isActive, ...others }) => {
       return <Button level="link" label={text} disabled={isActive} size="small" {...others} />;
@@ -28,7 +15,7 @@ export const Compact = (args) => (
   </Pagination>
 );
 
-Compact.args = {
+DefaultStory.args = {
   currentPage: 3,
   numPages: 21,
 };