+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Removed
 
-- `StatusBullet, TimerPulser, DatePickerRange`: ([@qubis741](https://github.com/qubis741) in [#1889](https://github.com/teamleadercrm/ui/pull/1889))
+- [BREAKING] `StatusBullet, TimerPulser, DatePickerRange`: ([@qubis741](https://github.com/qubis741) in [#1889](https://github.com/teamleadercrm/ui/pull/1889))
 
 ### Fixed