+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Removed
 
-- `LoadingMolecule`: ([@qubis741](https://github.com/qubis741) in [#1888](https://github.com/teamleadercrm/ui/pull/1888)) 
+- [BREAKING] `LoadingMolecule`: ([@qubis741](https://github.com/qubis741) in [#1888](https://github.com/teamleadercrm/ui/pull/1888)) 
 
 ### Fixed