+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Removed
 
+- [BREAKING] `ScrollContainer`: ([@qubis741](https://github.com/qubis741) in [#1891](https://github.com/teamleadercrm/ui/pull/1891))
+
 ### Fixed
 
 ### Dependency updates