+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 ### Dependency updates
 
+- Upgraded all dependencies to latest except `react`, `react-dom`, `react-day-picker` (latest is beta), `postcss-preset-env`, `postcss-each` ([@qubis741](https://github.com/qubis741) in [#1893](https://github.com/teamleadercrm/ui/pull/1893))
+
 ## [11.0.0] - 2021-12-20
 
 ### Added